feat: setup first classes for project
