--- src/test/java/br/com/portugol_reverso/MainTest.java
@@ -1,13 +1,0 @@
-package br.com.portugol_reverso;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-
-class MainTest {
-
-    @Test
-    void passWrongCommandLine() {
-        Assertions.assertThrows(RuntimeException.class, () -> Main.main("arquivo1.txt", "arquivo2.txt"));
-    }
-}
